[faet] 모달만 스토리북에 올리기 위해서 Modal 에서 ModalBox 를 분리
